Move ReputationChangeKind to network-api
